feat: move controllers and add OpenCV Implementation
